backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==4.2.21
django-filter==25.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flake8==7.2.0
flake8-isort==6.1.2
idna==3.10
isort==6.0.1
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.15.1
pillow==11.2.1
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
